refactor: address review — share auth-chain validation, fix stale comment
- socket-server now calls the shared validateAuthChain() instead of re-inlining
  the ownerAddress/parseEmphemeralPayload/validateSignature sequence, so it can't
  drift from the HTTP handlers. Moved auth-chain.ts from controllers/ to logic/
  so the socket (logic) no longer imports from controllers/ (layering).
- corrected the identity storage-expiration comment (said "15 minutes"; the code
  uses a 1 hour TTL).

The other two review notes were already resolved on this branch: the error
handler comes from @dcl/http-commons (returns no message on 500), and the
readJsonBody wrapper was replaced by parseJsonBody.
